- fixed bug trying to cast a real number with lexical_cast<int>
Former-commit-id: 12079fad9280390f28222fbafa741f9cc6446e94

+++ pwiz_aux/msrc/utility/vendor_api/thermo/RawFile.cpp
@@ -1353,7 +1353,7 @@ void RawFileImpl::parseInstrumentMethod()
             if (regex_match(line, what, defaultIsolationWidthRegex))
             {
                 int msLevel = lexical_cast<int>(what[1]);
-                int isolationWidth = lexical_cast<int>(what[2]);
+                double isolationWidth = lexical_cast<double>(what[2]);
                 defaultIsolationWidthBySegmentAndMsLevel[scanSegment][msLevel] = isolationWidth;
                 continue;
             }
